create bwpicture
I am led to believe that it was not your intention to modify the code with utilities and imports.  So rather than implement a dialog or option pane I have simply modified the color to black and white here, if you want however I can implement a dialog or menu so you can switch between color and black and white, etc. I was unable to use the included method due to the recursive nature of the for loops which allow one element to be drawn with multiple colors, but without user input it is impossible to designate between the two.